Board

+++ src/main/webapp/WEB-INF/templates/board/view.jsp
@@ -59,7 +59,18 @@
                     <fmt:message key="첨부파일"/>
                 </th>
                 <td>
-                    ${board.fileName}
+                    <c:if test="${board.attachFiles != null && !board.attachFiles.isEmpty()}">
+                        <ul class="download-items">
+                            <c:forEach var="item" items="${board.attachFiles}" varStatus="status">
+                                <li>
+                                    ${status.count})
+                                    <a href="<c:url value='/file/download' />/${item.seq}">
+                                            ${item.fileName}
+                                    </a>
+                                </li>
+                            </c:forEach>
+                        </ul>
+                    </c:if>
                 </td>
 
             </tr>
@@ -71,14 +82,28 @@
                     </a>
                 </div>
                 <div class="boardButton">
-                    <a href="${saveUrl}/${board.artNo}">
-                        <fmt:message key="수정하기" />
-                    </a>
+                    <c:if test="${member.userNo!=board.userNo}" >
+                        <a href="#" onclick="alert('작성자만 수정할 수 있습니다');">
+                            <fmt:message key="수정하기" />
+                        </a>
+                    </c:if>
+                    <c:if test="${member.userNo==board.userNo}" >
+                        <a href="${saveUrl}/${board.artNo}">
+                            <fmt:message key="수정하기" />
+                        </a>
+                    </c:if>
                 </div>
                 <div class="boardButton">
-                    <a href="${deleteUrl}/${board.artNo}">
+                    <c:if test="${member.userNo==board.userNo}" >
+                        <a href="${deleteUrl}/${board.artNo}" onclick="return confirm('게시글을 정말 삭제하시겠습니까')">
                         <fmt:message key="삭제하기" />
-                    </a>
+                        </a>
+                    </c:if>
+                    <c:if test="${member.userNo!=board.userNo}" >
+                        <a href="#" onclick="alert('작성자만 삭제할 수 있습니다');">
+                            <fmt:message key="삭제하기" />
+                        </a>
+                    </c:if>
                 </div>
             </div>
         </form>
